Demote cpdb-backend-cups to Recommends
If cpdb-backend-cups is installed but cups isn't,
network printers are still shown but can't be used.

Someone may want to uninstall CUPS if they don't want
to print using CUPS. Perhaps one day there will be
a CPDB backend for Cloud printing for instance.

At least in Ubuntu, we can't apt-depend on CUPS because
CUPS may be provided as a Snap soon.