d/rules: Ignore two reftests that fail on i386 only
In these tests, the reference rendering of "Hello world" is as one line,
but on i386 (possibly due to different rounding) it is word-wrapped onto
two lines.

I previously thought this was a regression between 4.10 and 4.12, but
it seems that instead, it was triggered by some other component being
upgraded, possibly fontconfig 2.14.2 or pango 1.51.0.