gsk/gl: cache Cairo textures
In previous episodes (cf. f44398f), we disabled a little piece of code I wrote that
badly cached Cairo render results to a GL texture. Conceptually the idea was good,
but we used the render node's address as the cache key, which made a lot of apps very
angry and was widely regarded as a bad move. You can read more about it in the
previously-cited commit's message.

Unfortunately for us, Cairo isn't getting any faster, so the need for such a cache
is alive and well. So here's what we do now: before rendering anything at all with
Cairo, we dump out the "thing" to render to a "Cairo script" (think PostScript).
We then hash that using an advanced & cryptographically secure function (CRC32),
and look in our texture cache for a texture with a matching hash. If we find one,
we skip the Cairo paint, we skip the RT allocation, and we skip the CPU->GPU copy.
If not... well, I will let you fill in the blanks there.

This, so far, seems to be working pretty well. Further testing is required.