d/rules: Build with -ffloat-store on i386
Some of the tests (and more in 4.15.x) fail if this is not done,
because i387 excess precision results in numeric values being calculated
differently. Because few GTK developers run it on i386 (and in
particular, upstream do not), running the test suite and not skipping
more of it than strictly necessary is important, because that's the only
mechanism available to discover whether the library has regressed
on less-used architectures.

This is probably detrimental to performance on i386, but if i386
performance was important to the project, we would likely be raising
the baseline so that we could use SSE2 instead of the i387.

Closes: #1077289