a11y: Fix A11Y groups
When an element is marked as a group, and is labelled by other
element, if an inner element gets the focus, the screen reader
must read the group label first and then the focused element
label.

This is how it worked in the previous version of GTK from Noble,
and in the one in Plucky; but the last update in Noble broke it,
and now, when there are nested elements, it doesn't work and the
screen reader doesn't read the upper levels, only the focused
element itself.

This bug was introduced with commit `4061e9ed92513bc1841f145ca026150fafe53310`,
and this patch just reverts the specific change in that commit
that breaks accessibility.

The reason is that the ORCA version in Noble doesn't know how to
handle ATSPI_ROLE_GROUPING, and that's why Gtk used previously
ATSPI_ROLE_PANEL instead. When ORCA added support for GROUPING,
GTK did the change too. Unfortunately, that broke A11Y in Noble.
Since both ATSPI_ROLE_PANEL and ATSPI_ROLE_GROUPING work in a
similar way, reverting the patch is safe.

Bug-Ubuntu: https://bugs.launchpad.net/ubuntu/+source/gtk4/+bug/2119005